DB & Updates

+++ Common/Devices/DigitalInput.cs
@@ -9,7 +9,7 @@
 
         public DigitalInput() : base() { }
 
-        public DigitalInput(string typeOfRegister, byte address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
+        public DigitalInput(string typeOfRegister, int address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
         {
             MaxValue = maxValue;
             MinValue = minValue;
