+++ Common/Devices/DigitalOutput.cs
@@ -9,7 +9,7 @@
 
         public DigitalOutput() : base() { }
 
-        public DigitalOutput(string typeOfRegister, byte address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
+        public DigitalOutput(string typeOfRegister, int address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
         {
             MaxValue = maxValue;
             MinValue = minValue;
