+++ Server/Models/DigitalInputModel.cs
@@ -17,7 +17,7 @@ namespace Server.Models
 
         public DigitalInputModel() : base() { }
 
-        public DigitalInputModel(string typeOfRegister, byte address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
+        public DigitalInputModel(string typeOfRegister, int address, string type, string description, byte maxValue, byte minValue, byte initialValue, byte value) : base(typeOfRegister, address, type, description)
         {
             MaxValue = maxValue;
             MinValue = minValue;
